+++ js/app.js
@@ -1,5 +1,4 @@
 import Vue from "vue";
-import VueI18Next from "@panter/vue-i18next";
 import App from "../src/App.vue";
 import store from "../src/app-store";
 import loadAddons from "../src/addons";
@@ -30,19 +29,18 @@ import MouseHoverPopupView from "../modules/mouseHover/view";
 import QuickHelpView from "../modules/quickHelp/view";
 import WindowView from "../modules/window/view";
 import SidebarView from "../modules/sidebar/view";
-import MeasureView from "../modules/tools/measure/view";
 import ShadowView from "../modules/tools/shadow/view";
 import ParcelSearchView from "../modules/tools/parcelSearch/view";
 import SearchByCoordView from "../modules/tools/searchByCoord/view";
 import LineView from "../modules/tools/pendler/lines/view";
 import AnimationView from "../modules/tools/pendler/animation/view";
 import FilterView from "../modules/tools/filter/view";
-import SaveSelectionView from "../modules/tools/saveSelection/view";
 import StyleWMSView from "../modules/tools/styleWMS/view";
 import StyleVTView from "../modules/tools/styleVT/view";
 import LayerSliderView from "../modules/tools/layerSlider/view";
 import CompareFeaturesView from "../modules/tools/compareFeatures/view";
 import RemoteInterfaceVue from "../src/plugins/remoteInterface/RemoteInterface";
+import {initiateVueI18Next} from "./vueI18Next";
 
 /**
  * WFSFeatureFilterView
@@ -98,7 +96,8 @@ async function loadApp () {
     const legacyAddons = Object.is(ADDONS, {}) ? {} : ADDONS,
         utilConfig = {},
         layerInformationModelSettings = {},
-        style = Radio.request("Util", "getUiStyle");
+        style = Radio.request("Util", "getUiStyle"),
+        vueI18Next = initiateVueI18Next();
     /* eslint-disable no-undef */
     let app = {},
         searchbarAttributes = {};
@@ -131,14 +130,12 @@ async function loadApp () {
 
     store.commit("setConfigJs", Config);
 
-    Vue.use(VueI18Next);
-
     app = new Vue({
         el: "#masterportal-root",
         name: "VueApp",
         render: h => h(App),
         store,
-        i18n: new VueI18Next(i18next, {namespaces: ["additional", "common"]})
+        i18n: vueI18Next
     });
 
 
@@ -219,10 +216,6 @@ async function loadApp () {
                 new ShadowView({model: tool});
                 break;
             }
-            case "measure": {
-                new MeasureView({model: tool});
-                break;
-            }
             case "print": {
                 /**
                  * PrintView2
@@ -245,10 +238,6 @@ async function loadApp () {
             }
             case "searchByCoord": {
                 new SearchByCoordView({model: tool});
-                break;
-            }
-            case "saveSelection": {
-                new SaveSelectionView({model: tool});
                 break;
             }
             /**
@@ -378,7 +367,7 @@ async function loadApp () {
 
     if (Config.addons !== undefined) {
         Radio.channel("Addons");
-        const i18nextLanguages = i18next && i18next.options.hasOwnProperty("getLanguages") ? i18next.options.getLanguages() : {};
+        const i18nextLanguages = vueI18Next && vueI18Next.options.hasOwnProperty("getLanguages") ? vueI18Next.options.getLanguages() : {};
         let initCounter = 0;
 
         Config.addons.forEach((addonKey) => {
@@ -398,7 +387,7 @@ async function loadApp () {
                         /* webpackInclude: /[\\\/]additional.json$/ */
                         `../addons/${addonKey}/locales/${lng}/additional.json`)
                         .then(({default: additionalLocales}) => {
-                            i18next.addResourceBundle(lng, "additional", additionalLocales, true);
+                            vueI18Next.addResourceBundle(lng, "additional", additionalLocales, true);
                             initCounter--;
                             checkInitCounter(initCounter, legacyAddons);
                         }).catch(error => {