+++ modules/core/util.js
@@ -44,7 +44,6 @@ const Util = Backbone.Model.extend(/** @lends Util.prototype */{
      * @listens Core#RadioTriggerUtilHideLoader
      * @listens Core#RadioTriggerUtilShowLoader
      * @listens Core#RadioTriggerUtilSetUiStyle
-     * @listens Core#RadioTriggerUtilCopyToClipboard
      * @listens Core#event:changeIsViewMobile
      * @fires Core#RadioTriggerIsViewMobileChanged
      * @fires Alerting#RadioTriggerAlertAlert
@@ -100,8 +99,7 @@ const Util = Backbone.Model.extend(/** @lends Util.prototype */{
             "hideLoader": this.hideLoader,
             "hideLoadingModule": this.hideLoadingModule,
             "showLoader": this.showLoader,
-            "setUiStyle": this.setUiStyle,
-            "copyToClipboard": this.copyToClipboard
+            "setUiStyle": this.setUiStyle
         }, this);
 
         // initial isMobileView setzen
@@ -448,52 +446,6 @@ const Util = Backbone.Model.extend(/** @lends Util.prototype */{
      */
     sortObjectsAsAddress: function (input) {
         return input.sort(this.sortAddress.bind(this));
-    },
-
-    /**
-     * Kopiert den Inhalt des Event-Buttons in die Zwischenablage, sofern der Browser das Kommando akzeptiert.
-     * behaviour of ios strange used solution from :
-     * https://stackoverflow.com/questions/34045777/copy-to-clipboard-using-javascript-in-ios
-     * @param  {el} el element to copy
-     * @fires Alerting#RadioTriggerAlertAlert
-     * @fires Alerting#RadioTriggerAlertAlert
-     * @returns {void}
-     */
-    copyToClipboard: function (el) {
-        const oldReadOnly = el.readOnly,
-            oldContentEditable = el.contentEditable,
-            range = document.createRange(),
-            selection = window.getSelection();
-
-        el.readOnly = false;
-        el.contentEditable = true;
-
-        range.selectNodeContents(el);
-        selection.removeAllRanges();
-        if (!this.isInternetExplorer()) {
-            selection.addRange(range);
-        }
-        el.setSelectionRange(0, 999999); // A big number, to cover anything that could be inside the element.
-
-        el.readOnly = oldReadOnly;
-        el.contentEditable = oldContentEditable;
-
-        try {
-            document.execCommand("copy");
-            Radio.trigger("Alert", "alert", {
-                text: i18next.t("common:modules.tools.saveSelection.contentSaved"),
-                kategorie: "alert-info",
-                position: "top-center",
-                fadeOut: 5000
-            });
-        }
-        catch (e) {
-            Radio.trigger("Alert", "alert", {
-                text: i18next.t("common:modules.tools.saveSelection.contenNotSaved"),
-                kategorie: "alert-info",
-                position: "top-center"
-            });
-        }
     },
 
     /**