+++ src/modules/tools/draw/store/stateDraw.js
@@ -95,12 +95,13 @@ const symbol = {
         deactivateGFI: true,
         download: {
             dataString: "",
-            enabled: false,
+            enabled: true,
             features: [],
             file: "",
             fileName: "",
             fileUrl: "",
             formats: ["KML", "GEOJSON", "GPX"], // NOTE(roehlipa): If this can't be configured, then it may be moved out of the state IMO.
+            preSelectedFormat: "",
             selectedFormat: ""
         },
         drawInteraction: null,