+++ src/modules/tools/indexTools.js
@@ -10,8 +10,10 @@ import actions from "./actionsTools";
 import Draw from "./draw/store/indexDraw";
 import FileImport from "./fileImport/store/indexFileImport";
 import Gfi from "./gfi/store/indexGfi";
+import SaveSelection from "./saveSelection/store/indexSaveSelection";
 import ScaleSwitcher from "./scaleSwitcher/store/indexScaleSwitcher";
 import SupplyCoord from "./supplyCoord/store/indexSupplyCoord";
+import Measure from "./measure/store/indexMeasure";
 
 /**
  * This is here to test app-store/utils/composeModules.
@@ -23,8 +25,10 @@ export default {
         Draw,
         FileImport,
         Gfi,
+        SaveSelection,
         ScaleSwitcher,
-        SupplyCoord
+        SupplyCoord,
+        Measure
     },
     state,
     getters,