+++ CHANGELOG.md
@@ -12,16 +12,21 @@
 - On tools Imported Layer in theme tree integrated function under Addons
 - On tools Geometrie/Gebäude Analyse funtion on tools under Addons
 - On tools Taktische Zeichen on tools under Addons
+
 ### Changed
 - The language of the changelog has been changed to English.
+- The measure tool has been migrated to Vue.
 - Migrated the Download Tool from Backbone.js to Vue.js and integrated it into the Draw Tool.
+- Migrated the SaveSelection Tool from Backbone.js to Vue.js. It also no longer has a Radio channel.
+- copyToClipboard is now a reusable action and can no longer be triggered through the Radio.
+
 ### Deprecated
+- Deprecated the parameter `simpleMap` in the config.js and moved its configuration to the `saveSelection` tool in the config.json.
 
 ### Removed
 
 ### Fixed
 - If the Save selection tool is not configured, the button for it is no longer displayed in the topic tree.
-
 
 ---